ODROID-M1: Add 'oem2env' command
'oem2env' command is introduced to help ODROID-M1 users who want to
provide vendor's S/N that is stored in the OTP area. The data in the area
will be treated as non-string type after reading to memory, it cannot be
set to an environment variable. Therefore 'oem2env' read 16 bytes in OTP
area that is allocated for OEM users and set to a given environment
variable.

    > oem2env 1 oem-serial

OEM data (16 bytes) in OTP area will be stored as 'oem-serial' and this
can be set to device tree using a command,

    > fdt set /chosen oem-serial ${oem-serial}

then is will be presented in Linux system,

    $ cat /proc/device-tree/chosen/oem-serial
    HK2024-ODROID-M1

Requested thread:
    https://forum.odroid.com/viewtopic.php?f=212&t=48501

Signed-off-by: Dongjin Kim <tobetter@gmail.com>
Change-Id: I9daf534e9807a5e2e4bc4eb892130f3b588c9fa4